Test Automation Concepts/Archive/pytest-selenium-starter/requirements.txt
selenium>=4.23.0
pytest>=8.2.0
pytest-html>=4.1.1
pytest-xdist>=3.6.1
pyyaml>=6.0.1
